+++ src/store/index.ts
@@ -1,7 +1,7 @@
 import { configureStore, combineReducers } from '@reduxjs/toolkit';
 import bookmarks from './modules/bookmarksSlice';
 import tabs from './modules/tabsSlice';
-import interfaces from './modules/intefaceSlice';
+import interfaces from './modules/windowsSlice';
 import logger from 'redux-logger';
 
 const reducer = combineReducers({ bookmarks, tabs, interfaces });
