+++ src/store/modules/windowsSlice.ts
